+++ Rest/src/main/java/sber/spring/Rest/service/ClientService.java
@@ -21,23 +21,23 @@ public class ClientService {
     @Autowired
     private BinRepository binRepository;
 
-    public long saveClient(Client client){
-      if (!clientRepository.isClient(client)){
-          return clientRepository.signClient(client, binRepository.createBin());
-      } else {
-          return 0;
-      }
+    public int saveClient(Client client) {
+        if (!clientRepository.isClient(client)) {
+            return clientRepository.signClient(client, binRepository.createBin());
+        } else {
+            return 0;
+        }
     }
 
-    public void updateBin(long clientId){
+    public void updateBin(int clientId){
         clientRepository.updateBin(clientId, binRepository.getBin(clientId));
     }
 
-    public Optional<Client> searchClientRep(long id) {
+    public Optional<Client> searchClientRep(int id) {
         return clientRepository.searchClient(id);
     }
 
-    public boolean deleteClientFromRep(long id) {
+    public boolean deleteClientFromRep(int id) {
         return clientRepository.deleteClient(id);
     }
 
@@ -74,22 +74,22 @@ public class ClientService {
         return binRepository.isInBin(binId, productId);
     }
 
-    public void addToBin(long binId, int productId) {
+    public void addToBin(int binId, int productId) {
         binRepository.add(binId, productRepository.search(productId).get());
         clientRepository.updateBin(binId, binRepository.getBin(binId));
     }
 
-    public void changeQuantity(long binId, long productId, int quantity) {
+    public void changeQuantity(int binId, int productId, int quantity) {
         binRepository.changeQuantity(binId, productId, quantity);
         clientRepository.updateBin(binId, binRepository.getBin(binId));
     }
 
-    public void deleteProductFromBin(long binId, long productId) {
+    public void deleteProductFromBin(int binId, int productId) {
         binRepository.deleteFromBin(binId, productId);
         clientRepository.updateBin(binId, binRepository.getBin(binId));
     }
 
-    public void pay(long binId) {
+    public void pay(int binId) {
         binRepository.pay(binId);
         clientRepository.updateBin(binId, binRepository.getBin(binId));
     }
